:recycle: :fire: Remove create() and associated types and methods from vote_transaction
