+++ packages/lisk-transactions/src/transactions/3_vote_transaction.ts
@@ -17,16 +17,10 @@ import * as BigNum from 'browserify-bignum';
 import { VOTE_FEE } from '../constants';
 import { TransactionError, TransactionMultiError } from '../errors';
 import { Account, Status, TransactionJSON } from '../transaction_types';
-import {
-	createBaseTransaction,
-	CreateBaseTransactionInput,
-	prependMinusToPublicKeys,
-	prependPlusToPublicKeys,
-} from '../utils';
+import { CreateBaseTransactionInput } from '../utils';
 import {
 	isTypedObjectArrayWithKeys,
 	validateAddress,
-	validatePublicKeys,
 	validator,
 } from '../utils/validation';
 import {
@@ -92,23 +86,6 @@ export const voteAssetFormatSchema = {
 	},
 };
 
-const validateInputs = ({
-	votes = [],
-	unvotes = [],
-}: CreateVoteAssetInput): void => {
-	if (!Array.isArray(votes)) {
-		throw new Error(
-			'Please provide a valid votes value. Expected an array if present.',
-		);
-	}
-	if (!Array.isArray(unvotes)) {
-		throw new Error(
-			'Please provide a valid unvotes value. Expected an array if present.',
-		);
-	}
-	validatePublicKeys([...votes, ...unvotes]);
-};
-
 export class VoteTransaction extends BaseTransaction {
 	public readonly containsUniqueData = true;
 	public readonly asset: VoteAsset;
@@ -131,45 +108,6 @@ export class VoteTransaction extends BaseTransaction {
 		}
 		this.asset = tx.asset as VoteAsset;
 		this._fee = new BigNum(VOTE_FEE);
-	}
-
-	public static create(input: CastVoteInput): object {
-		validateInputs(input);
-		const { passphrase, secondPassphrase, votes = [], unvotes = [] } = input;
-
-		const plusPrependedVotes = prependPlusToPublicKeys(votes);
-		const minusPrependedUnvotes = prependMinusToPublicKeys(unvotes);
-		const allVotes: ReadonlyArray<string> = [
-			...plusPrependedVotes,
-			...minusPrependedUnvotes,
-		];
-
-		const transaction = {
-			...createBaseTransaction(input),
-			type: 3,
-			fee: VOTE_FEE.toString(),
-			asset: {
-				votes: allVotes,
-			},
-		};
-
-		if (!passphrase) {
-			return transaction;
-		}
-
-		const transactionWithSenderInfo = {
-			...transaction,
-			// SenderId and SenderPublicKey are expected to be exist from base transaction
-			senderId: transaction.senderId as string,
-			senderPublicKey: transaction.senderPublicKey as string,
-			recipientId: transaction.senderId as string,
-			recipientPublicKey: transaction.senderPublicKey,
-		};
-
-		const voteTransaction = new VoteTransaction(transactionWithSenderInfo);
-		voteTransaction.sign(passphrase, secondPassphrase);
-
-		return voteTransaction.toJSON();
 	}
 
 	protected getAssetBytes(): Buffer {
